separate notifications for each complete offline file download

+++ src/main/java/org/amahi/anywhere/util/Downloader.java
@@ -235,10 +235,10 @@ public class Downloader extends BroadcastReceiver {
                 Cursor cursor = getDownloadManager(context).query(q);
                 if (cursor != null && cursor.moveToFirst()) {
                     cursor.moveToFirst();
-                    int bytes_downloaded = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_BYTES_DOWNLOADED_SO_FAR));
-                    int bytes_total = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_TOTAL_SIZE_BYTES));
-                    int progress = bytes_total != 0 ? 100 * bytes_downloaded / bytes_total : 0;
-                    downloadCallbacks.downloadProgress((int) id, progress);
+                    long bytes_downloaded = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_BYTES_DOWNLOADED_SO_FAR));
+                    long bytes_total = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_TOTAL_SIZE_BYTES));
+                    long progress = bytes_total != 0 ? (100 * bytes_downloaded) / bytes_total : 0;
+                    downloadCallbacks.downloadProgress((int) id, (int) progress);
 
                     int status = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_STATUS));
                     switch (status) {
@@ -246,7 +246,7 @@ public class Downloader extends BroadcastReceiver {
                         case DownloadManager.STATUS_PAUSED:
                             if (!(isNetworkAvailable())) {
                                 isDownloading = false;
-                                downloadCallbacks.downloadPaused(id, progress);
+                                downloadCallbacks.downloadPaused(id, (int) progress);
                             } else {
                                 downloadCallbacks.downloadError(id);
                                 isDownloading = false;
